pyproject.toml
[project]
name = "physoce-datasets"
version = "0.1.0"
description = "A Python package and CLI for downloading various physical oceanographic datasets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cdsapi>=0.7.7",
    "cfgrib>=0.9.15.1",
    "click>=8.2.1",
    "copernicusmarine>=2.3.0",
    "ecmwf-datastores-client>=0.5.1",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "xarray[accel,io,parallel]>=2026.2.0",
]

[dependency-groups]
notebook = ["ipykernel>=7.2.0", "matplotlib>=3.10.8"]
